==> 11_eval_new.ipynb <==
# New valuation for realistic run with morphology
1. Create a set of mentions from every sentence: sent_pred_set, sent_true_set
   1. add numbering to mentions if they are the same
1. Join sets with sentence identifer: pred_set, true_set
1. Precision: intersection(pred_set, true_set) / true_set
1. Recall: intersection(pred_set, true_set) / pred_set

## Prepare corpus

%load_ext autoreload
%autoreload 2

%matplotlib inline

import pandas as pd
import numpy as np

import matplotlib.pyplot as plt
import seaborn as sns
sns.set_context('paper')
sns.set_style('white')

from conlleval import evaluate

df = (pd.read_csv('curation.csv.gz')
      .assign(sent=lambda x: x.file+'_'+x.sent.astype(str)))

df.head()

df.head().T

x=df
sents = x.groupby('sent')[['token', 'biose', 'ner_type']].apply(lambda x: x.values.tolist())
sents[:5]

## Create set of mentions from every sentence

sents_fixed = list(zip(list(sents.index), [[(tok, bio[0], cat) for tok, bio, cat in sent] for sent in sents]))

sents_fixed[:2]

from collections import defaultdict
from itertools import islice

def sent_to_mentions_dict(sent, sent_id, truncate=80):
    mentions = defaultdict(lambda: 0)
    current_mention= None
    current_cat = None
    for tok, bio, cat in islice(sent, truncate):
        if bio=='S':
            mentions[(sent_id, tok, cat)]+=1
        if bio=='B':
            current_mention = [tok]
            current_cat = cat
        if bio=='I' and current_mention is not None:
            current_mention.append(tok)
        if bio=='E' and current_mention is not None:
            current_mention.append(tok)
            mentions[(sent_id, ' '.join(current_mention), current_cat)]+=1
        if bio=='O':
            current_mention = None
            current_cat = None
    return mentions
ments = [sent_to_mentions_dict(sent, sent_id) for sent_id, sent in sents_fixed]
        

ments[:5]

def get_ment_set(ments):
    ment_set = []
    for ment in ments:
        for k, val in ment.items():
            for i in range(val):
                ment_set.append((k[0], k[1], k[2], i+1))
    return ment_set
ment_set = ment_set[:5]
            

def sents_to_mentions(sents):
    sents_fixed = zip(list(sents.index), [[(tok, bio[0], cat) for tok, bio, cat in sent] for sent in sents])
    ments = [sent_to_mentions_dict(sent, sent_id) for sent_id, sent in sents_fixed]
    ment_set = get_ment_set(ments)
    return ment_set

import pickle
with open('treebank_results_2_test25.pkl', 'rb') as f:
    tb_results = pickle.load(f)
len(tb_results)

all_preds = [r[2] for r in tb_results]
all_preds[0][0][:20]

from sklearn.model_selection import train_test_split

splits = [train_test_split(range(len(sents)), 
                           sents, test_size=0.25, 
                           random_state=42+i) 
          for i in range(3)]

sum([len(s)-80 for s in splits[0][3] if len(s)>80])

sum(len(s) for s in splits[0][3]), len(all_preds[0][0])

def get_sents_with_pred_tags(splits, preds, truncate=80):
    sents_preds = []
    for split, pred in zip(splits, preds):
        spl_preds = []
        test_sents = split[3]
        
        i=0
        for sent in test_sents:
            new_sent = []
            for tok, bio, cat in islice(sent, truncate):
                pred_tag = pred[i].split('-')
                pred_bio = pred_tag[0]
                if len(pred_tag)>1:
                    pred_cat = pred_tag[1]
                else:
                    pred_cat = '_'
                new_sent.append((tok, pred_bio, pred_cat))
                i+=1
            spl_preds.append(new_sent)
        spl_preds = pd.Series(spl_preds, index=test_sents.index)
        sents_preds.append(spl_preds)
    return sents_preds

all_sents_preds = [get_sents_with_pred_tags(splits, preds) for preds in all_preds]

splits_sents = [sents.iloc[split[1]] for split in splits]
splits_ments = [sents_to_mentions(ss) for ss in splits_sents]
all_ments_preds = [[sents_to_mentions(sp) for sp in sents_preds] for sents_preds in all_sents_preds]

splits_ments[0][:10]


all_ments_preds[0][0][:10]

def evaluate_mentions(true_ments, pred_ments):
        t, p = set(true_ments), set(pred_ments)
        correct = p.intersection(t)
        prec = len(correct) / len(t)
        recall = len(correct) / len(p)
        f1 = 2*prec*recall/(prec+recall)
        print('Precision:', round(prec, 2))
        print('Recall:   ', round(recall, 2))
        print('F1:       ', round(f1, 2))
        print('FP ex.:', [e[1] for e in list(p-t)[:5]])
        print('FN ex.:', [e[1] for e in list(t-p)[:5]])
        return prec, recall, f1
for ments_preds in all_ments_preds:
    for t, p in zip(splits_ments, ments_preds):
        print(len(t), 'mentions,', len(p), 'found,', len(set(p).intersection(set(t))), 'correct.')
        evaluate_mentions(t, p)

## Ru
