==> 18_treebank_token_dfs.ipynb <==
import pandas as pd
from conllu import parse


## Read spmrl from conllu to DF

from collections import OrderedDict

def get_conllu_df(path, remove_duplicates=False, remove_very_similar=False):
    with open(path, 'r', encoding='utf8') as f:
        sp_conllu = parse(f.read())
    fixed = []
    dup_to_remove = set()
    very_sim_to_remove = set()
    for tl in sp_conllu:
        if (remove_duplicates and int(tl.metadata['sent_id']) in dup_to_remove 
            or remove_very_similar and int(tl.metadata['sent_id']) in very_sim_to_remove):
            print ('skipped', tl.metadata['sent_id'])
            continue
        for tok in tl:
            t = OrderedDict(tok)
            if type(t['id']) is not tuple:
                if t['feats'] is not None:
                    t.update({'feats_'+f: v for f, v in t['feats'].items()})
                del(t['feats'])
                if t['misc'] is not None:
                    t.update({'misc_'+f: v for f, v in t['misc'].items()})
                del(t['misc'])
                t.update(tl.metadata)
                fixed.append(t)
            if remove_duplicates:
                dup_to_remove = dup_to_remove | set(eval(tl.metadata['duplicate_sent_id']))
            if remove_very_similar:
                very_sim_to_remove = dup_to_remove | set(eval(tl.metadata['very_similar_sent_id']))

    df = (pd.DataFrame(fixed)
          .assign(sent_id = lambda x: x.sent_id.astype(int))
          .assign(global_sent_id = lambda x: x.global_sent_id.astype(int))
          .assign(misc_token_id = lambda x: x.misc_token_id.astype(int))

         )
    return df

        
spdf = get_conllu_df('align/spmrl_fixed.conllu', remove_duplicates=True)
    

spdf[['biose_only', 'ner_type']] = spdf.misc_biose.str.split('-', expand=True)

spdf.head().T

spdf.groupby('set').sent_id.nunique()

spdf.dtypes

spdf.ner_type.value_counts()

(spdf
          .groupby(['sent_id', 'misc_token_id', 'misc_token_str']).apply(lambda x: ','.join(x.biose_only.tolist())))

## Create token DF

def get_token_biose(df):
    def _single_token_conversion(tok):
        all_bio = tok.biose_only.tolist()
        all_typ = set(tok.ner_type.dropna().tolist())
        if len(all_typ)>1:
            return 'O'
        if 'S' in all_bio:
            new_bio = 'S'
        elif 'B' in all_bio and 'E' in all_bio:
            new_bio = 'S'
        elif 'B' in all_bio:
            new_bio = 'B'
        elif 'E' in all_bio:
            new_bio = 'E'
        elif 'I' in all_bio:
            new_bio = 'I'
        else:
            return 'O'
        return new_bio+'-'+all_typ.pop()
    
    df = (spdf
          .groupby(['sent_id', 'misc_token_id', 'misc_token_str'])
          .apply(_single_token_conversion)
          .reset_index().rename(columns={0:'biose'})
         )
    return df
    
tok_ner = get_token_biose(spdf)

tok_ner.head()

tok_ner.head(30)

spdf.columns[spdf.columns.str.startswith('feats')]

from tqdm import tqdm
# Create and register a new `tqdm` instance with `pandas`
# (can use tqdm_gui, optional kwargs, etc.)
tqdm.pandas()

#**{'upostag': ','.join(x.upostag.tolist())}
features = (spdf
            .groupby(['sent_id', 'misc_token_id', 'misc_token_str'])
            .progress_apply(lambda x: pd.Series({f: '^'.join(x[f].fillna('').tolist()) for f in spdf.columns[spdf.columns.str.startswith('feats')]}
            ))
           )

features.head(50)

upos_form_lemma_deprel = (spdf
            .groupby(['sent_id', 'misc_token_id', 'misc_token_str'])
            .progress_apply(lambda x: pd.Series({'upostag': '^'.join(x.upostag.tolist()),
                                                 'form': '^'.join(x.form.tolist()),
                                                 'lemma': '^'.join(x.lemma.tolist()),
                                                 'deprel': '^'.join(x.deprel.tolist()),}
            ))
           )

tokens_ner_with_upos_feats = pd.concat([tok_ner.set_index(['sent_id', 'misc_token_id', 'misc_token_str']), upos_form_lemma_deprel, features], axis=1)
tokens_ner_with_upos_feats = (tokens_ner_with_upos_feats
                              .reset_index()
                              .set_index(['sent_id', 'misc_token_id', 'misc_token_str'])
                              .sort_index()
                              .assign(set = lambda x: x.index.get_level_values('sent_id').map(spdf[['sent_id', 'set']].drop_duplicates().set_index('sent_id')['set']))
                              .reset_index()
                             )
tokens_ner_with_upos_feats.head()


tokens_ner_with_upos_feats.deprel.value_counts()

## Run YAP on tokens

tn = tokens_ner_with_upos_feats

import os
yap_output_dir = 'new_yap_outputs'

os.mkdir(yap_output_dir)

def get_token_output(s):
    return '\n'.join(s.misc_token_str.tolist())+'\n\n'

def write_tokens_to_file(tok_df, corp_set, output_path):
    with open(output_path, 'w') as of:
        for sent in tok_df[tok_df.set==corp_set].reset_index().groupby('sent_id').apply(get_token_output):
            of.write(sent)


write_tokens_to_file(tn, 'dev', os.path.join(yap_output_dir, 'spmrl_dev_tokens.txt'))
write_tokens_to_file(tn, 'test', os.path.join(yap_output_dir, 'spmrl_test_tokens.txt'))
write_tokens_to_file(tn, 'train', os.path.join(yap_output_dir, 'spmrl_train_tokens.txt'))

yap_path = '/home/dan/yapproj/src/yap/yap'

!export GOPATH=/home/dan/yapproj

!{yap_path}

!{yap_path} hebma -raw new_yap_outputs/spmrl_dev_tokens.txt -out new_yap_outputs/spmrl_dev.lattices

!{yap_path} joint -in new_yap_outputs/spmrl_dev.lattices -os new_yap_outputs/spmrl_dev_seg.conll -om new_yap_outputs/spmrl_dev_map.conll -oc new_yap_outputs/spmrl_dev_dep.conll

!{yap_path} hebma -raw new_yap_outputs/spmrl_test_tokens.txt -out new_yap_outputs/spmrl_test.lattices

!{yap_path} joint -in new_yap_outputs/spmrl_test.lattices -os new_yap_outputs/spmrl_test_seg.conll -om new_yap_outputs/spmrl_test_map.conll -oc new_yap_outputs/spmrl_test_dep.conll

## Create token DF from YAP outputs

import os
import numpy as np

def make_conll_df(path, add_head_stuff=False):
    # CoNLL file is tab delimeted with no quoting
    # quoting=3 is csv.QUOTE_NONE
    df = (pd.read_csv(path, sep='\t', header=None, quoting=3, comment='#',
                names = ['id', 'form', 'lemma', 'upostag', 'xpostag', 'feats', 'head', 'deprel', 'deps', 'misc'])
                # add sentence labels
                .assign(sent = lambda x: (x.id==1).cumsum())
                # replace bad root dependency tags
                .replace({'DEPREL': {'prd': 'ROOT'}})
               )
    
    if add_head_stuff:
        df = df.merge(df[['ID', 'FORM', 'sent', 'UPOS']].rename(index=str, columns={'FORM': 'head_form', 'UPOS': 'head_upos'}).set_index(['sent', 'ID']),
               left_on=['sent', 'HEAD'], right_index=True, how='left')
    return df

def read_lattices(path):
    df = (pd.read_csv(path, sep='\t', header=None, quoting=3, 
                names = ['ID1', 'ID2', 'form', 'lemma', 'upostag', 'xpostag', 'feats', 'misc_token_id'])
                # add sentence labels
                .assign(sent = lambda x: (x.ID1==0).cumsum())
               )
    return df

flatten = lambda l: [item for sublist in l for item in sublist]

def get_feats(s):
    if s!='_' and s is not None and s is not np.nan:
        feats = OrderedDict()
        for f in s.split('|'):
            k,v = f.split('=')
            k='feats_'+k
            if k not in feats:
                feats[k] = v
            else:
                feats[k] = feats[k]+','+v
        return pd.Series(feats)
    else:
        return pd.Series()

def get_yap_output_df(tokens_path, dep_path, map_path):
    tokens = dict(flatten([[(str(j+1)+'_'+str(i+1), tok) for i, tok in enumerate(sent.split('\n'))]
              for j, sent in 
              enumerate(open(os.path.join(yap_output_dir, tokens_path), 'r').read().split('\n\n'))]))
    lattices = read_lattices(map_path)
    dep = make_conll_df(dep_path)
    df = (pd.concat([dep, lattices.misc_token_id], axis=1)
          .assign(sent_tok = lambda x: x.sent.astype(str) + '_' + x.misc_token_id.astype(str))
          .assign(misc_token_str = lambda x: x.sent_tok.map(tokens))
          .drop('sent_tok', axis=1)
          )
    df = pd.concat([df, df.feats.apply(get_feats)], axis=1).drop('feats', axis=1)
    return df


yap_dev = get_yap_output_df('spmrl_dev_tokens.txt', 'new_yap_outputs/spmrl_dev_dep.conll', 'new_yap_outputs/spmrl_dev_map.conll')
yap_test = get_yap_output_df('spmrl_test_tokens.txt', 'new_yap_outputs/spmrl_test_dep.conll', 'new_yap_outputs/spmrl_test_map.conll')


yap_dev.head()

#**{'upostag': ','.join(x.upostag.tolist())}
yap_dev_feats = (yap_dev
            .groupby(['sent', 'misc_token_id', 'misc_token_str'])
            .apply(lambda x: pd.Series({f: '^'.join(x[f].fillna('').tolist()) for f in yap_dev.columns[yap_dev.columns.str.startswith('feats')]}
            ))
           )
yap_dev_tok = (yap_dev
            .groupby(['sent', 'misc_token_id', 'misc_token_str'])
            .apply(lambda x: pd.Series({ 'upostag': '^'.join(x.upostag.fillna('').tolist()),
                                         'form': '^'.join(x.form.fillna('').tolist()),
                                         'lemma': '^'.join(x.lemma.fillna('').tolist()),
                                         'deprel': '^'.join(x.deprel.fillna('').tolist()),}
            ))
           )
yap_dev_tok = pd.concat([yap_dev_tok,yap_dev_feats], axis=1)

yap_test_feats = (yap_test
            .groupby(['sent', 'misc_token_id', 'misc_token_str'])
            .apply(lambda x: pd.Series({f: ','.join(x[f].fillna('').tolist()) for f in yap_test.columns[yap_test.columns.str.startswith('feats')]}
            ))
           )
yap_test_tok = (yap_test
            .groupby(['sent', 'misc_token_id', 'misc_token_str'])
            .apply(lambda x: pd.Series({ 'upostag': '^'.join(x.upostag.fillna('').tolist()),
                                         'form': '^'.join(x.form.fillna('').tolist()),
                                         'lemma': '^'.join(x.lemma.fillna('').tolist()),
                                         'deprel': '^'.join(x.deprel.fillna('').tolist()),}
            ))
           )
yap_test_tok = pd.concat([yap_test_tok, yap_test_feats], axis=1)

dev_sent_id_map = tn[tn.set=='dev'].sent_id.drop_duplicates().reset_index().drop('index', axis=1)['sent_id']
dev_sent_id_map.index = dev_sent_id_map.index+1
test_sent_id_map = tn[tn.set=='test'].sent_id.drop_duplicates().reset_index().drop('index', axis=1)['sent_id']
test_sent_id_map.index = test_sent_id_map.index+1
yap_dev_tok = (yap_dev_tok
               .reset_index()
               .assign(sent_id = lambda x: x['sent'].map(dev_sent_id_map)))
yap_test_tok = (yap_test_tok.reset_index()
               .assign(sent_id = lambda x: x.sent.map(test_sent_id_map)))

yap_dev['sent_id'] = yap_dev.sent.map(dev_sent_id_map)
yap_test['sent_id'] = yap_test.sent.map(test_sent_id_map)

yap_dev.to_csv('new_yap_outputs/yap_dev.csv.gz', index=False, compression='gzip')
yap_test.to_csv('new_yap_outputs/yap_test.csv.gz', index=False, compression='gzip')

yap_dev_tok.to_csv('new_yap_outputs/yap_dev_tokens.csv.gz', index=False, compression='gzip')
yap_test_tok.to_csv('new_yap_outputs/yap_test_tokens.csv.gz', index=False, compression='gzip')

tokens_ner_with_upos_feats.reset_index().to_csv('align/spmrl_tokens_df.csv.gz', index=False, compression='gzip')

tokens_ner_with_upos_feats
