==> 23_ner_layers_transforms.ipynb <==
import numpy as np
import pandas as pd

from ner_transforms import *

ud = pd.read_csv('align/uddf_fixed.csv.gz')
biose = get_all_layers_biose(ud)

biose.head(30)

biose.shape

ud.shape

ud = pd.concat([ud, biose.drop(['sent_id', 'id'], axis=1)], axis=1)
ud.shape

ud.head().T

sp = pd.read_csv('align/spdf_fixed.csv.gz', low_memory=False)

biose_sp = get_all_layers_biose(sp)

sp.shape, biose_sp.shape

sp = pd.concat([sp, biose_sp.drop(['sent_id', 'id'], axis=1)], axis=1)
sp.shape

sp[sp.biose!=sp.biose_layer0].T

sp.loc[sp.sent_id==1408, ['form', 'biose', 'biose_layer0', 'biose_layer1']]

sp.to_csv('align/spdf_fixed.csv.gz', compression='gzip', index=False)
ud.to_csv('align/uddf_fixed.csv.gz', compression='gzip', index=False)


sp[sp.biose!=sp.biose_layer0].T

sp.loc[sp.sent_id==4121, ['form', 'biose', 'biose_layer0', 'biose_layer1']]

ud[ud.biose!=ud.biose_layer0].T
