==> 31_data_for_ncrf.ipynb <==
# Create Data for NCRF

%load_ext autoreload
%autoreload 2

%matplotlib inline

import pandas as pd
import numpy as np

import matplotlib.pyplot as plt
import seaborn as sns
sns.set_context('paper')
sns.set_style('white')

from conlleval import evaluate

from ner_utils import *

import bclm

dropped = [5438, 5444, 5445, 5446, 5448, 5449, 5450, 5451, 5453, 5459]

spdf = bclm.read_dataframe('spmrl')

spdf.groupby('set').sent_id.agg([min, max])

spdf.groupby('set').sent_id.nunique()

yap_dev = bclm.read_yap_output('dev')
yap_test = bclm.read_yap_output('test')


yap_test.sent_id.nunique()

spdf = spdf[(~spdf.sent_id.isin(dropped))]


spdf.head()

yap_dev.head()

dev_tok_path = 'new_yap_outputs/spmrl_dev_tokens.txt'
test_tok_path = 'new_yap_outputs/spmrl_test_tokens.txt'

yap_pruned_all_dev   = bclm.read_yap_output(treebank_set=None, tokens_path=dev_tok_path, dep_path='data/pruned/dev_pruned_all_keep.conll', map_path='data/pruned/dev_pruned_all_keep.map')
yap_pruned_nono_dev  = bclm.read_yap_output(treebank_set=None, tokens_path=dev_tok_path,  dep_path='data/pruned/dev_pruned_non_o_keep.conll',  map_path='data/pruned/dev_pruned_non_o_keep.map')


## Morpheme Sentences

def get_sent_lists(df, group='sent_id', fields=['form', 'upostag', 'biose_layer0']):
    return df.groupby(group)[fields].apply(lambda x: x.values.tolist())

import os
out_folder = 'data/for_ncrf/'

def write_ncrf_file(sents, file_name, keep_pos=False):
    if keep_pos:
        file_name += '_pos'
    with open(os.path.join(out_folder, file_name+'.bmes'), 'w', encoding='utf8') as of:
        for sent in sents:
            for fields in sent: 
                new_pos = '[POS]'+fields[1]
                line = fields[0]
                if keep_pos:
                    line += ' ' + new_pos
                if len(fields)==3:
                    line += ' ' + fields[2]
                of.write(line+'\n')
            of.write('\n')
                    
        

yap_dev_sents =  get_sent_lists(yap_dev, fields=['form', 'upostag'])
yap_dev_sents.head()

write_ncrf_file(yap_dev_sents, 'morph_yap_dev', keep_pos=True)
write_ncrf_file(yap_dev_sents, 'morph_yap_dev')

yap_test_sents =  get_sent_lists(yap_test, fields=['form', 'upostag'])
yap_test_sents.head()

write_ncrf_file(yap_test_sents, 'morph_yap_test', keep_pos=True)
write_ncrf_file(yap_test_sents, 'morph_yap_test')

yap_pruned_all_dev =  get_sent_lists(yap_pruned_all_dev, fields=['form', 'upostag'])
yap_pruned_all_dev.head()

write_ncrf_file(yap_dev_sents, 'morph_yap_dev_pruned_all', keep_pos=True)
write_ncrf_file(yap_dev_sents, 'morph_yap_dev_pruned_all')

yap_pruned_nono_dev =  get_sent_lists(yap_pruned_nono_dev, fields=['form', 'upostag'])
yap_pruned_nono_dev.head()

write_ncrf_file(yap_dev_sents, 'morph_yap_dev_pruned_non_o', keep_pos=True)
write_ncrf_file(yap_dev_sents, 'morph_yap_dev_pruned_non_o')

gold_dev  = spdf[spdf.set=='dev']
gold_train = spdf[spdf.set=='train']
gold_test = spdf[spdf.set=='test']

yap_test.sent_id.nunique(), gold_test.sent_id.nunique(), 

gold_dev_sents = get_sent_lists(gold_dev)
gold_train_sents = get_sent_lists(gold_train)
gold_test_sents =  get_sent_lists(gold_test)
gold_test_sents.head()

write_ncrf_file(gold_dev_sents,   'morph_gold_dev', keep_pos=True)
write_ncrf_file(gold_dev_sents,   'morph_gold_dev')
write_ncrf_file(gold_train_sents, 'morph_gold_train', keep_pos=True)
write_ncrf_file(gold_train_sents, 'morph_gold_train')
write_ncrf_file(gold_test_sents, 'morph_gold_test', keep_pos=True)
write_ncrf_file(gold_test_sents, 'morph_gold_test')

## Concat BIOSE Sentences

biose_concat = bclm.get_token_df(spdf, fields = ['biose_layer0', 'upostag'])
biose_concat.head()

token_gold_dev_concat  = biose_concat[biose_concat.set=='dev']
token_gold_train_concat = biose_concat[biose_concat.set=='train']
token_gold_test_concat = biose_concat[biose_concat.set=='test']

token_gold_dev_concat_sents    = get_sent_lists(token_gold_dev_concat  , fields=['token_str', 'upostag', 'biose_layer0'])
token_gold_train_concat_sents  = get_sent_lists(token_gold_train_concat, fields=['token_str', 'upostag', 'biose_layer0'])
token_gold_test_concat_sents   = get_sent_lists(token_gold_test_concat , fields=['token_str', 'upostag', 'biose_layer0'])
token_gold_test_concat_sents.head()

write_ncrf_file(token_gold_dev_concat_sents, 'token_gold_dev_concat', keep_pos=True)
write_ncrf_file(token_gold_dev_concat_sents, 'token_gold_dev_concat')
write_ncrf_file(token_gold_train_concat_sents, 'token_gold_train_concat', keep_pos=True)
write_ncrf_file(token_gold_train_concat_sents, 'token_gold_train_concat')
write_ncrf_file(token_gold_test_concat_sents, 'token_gold_test_concat', keep_pos=True)
write_ncrf_file(token_gold_test_concat_sents, 'token_gold_test_concat')

## Fix BIOSE Sentences

biose_fix = bclm.get_token_df(spdf, fields = ['upostag'], biose=['biose_layer0'])
biose_fix.head()

token_gold_dev_fix  =     biose_fix[biose_fix.set=='dev']
token_gold_train_fix = biose_fix[biose_fix.set=='train']
token_gold_test_fix =  biose_fix[biose_fix.set=='test']

token_gold_dev_fix_sents    = get_sent_lists(token_gold_dev_fix  , fields=['token_str', 'upostag', 'biose_layer0'])
token_gold_train_fix_sents  = get_sent_lists(token_gold_train_fix, fields=['token_str', 'upostag', 'biose_layer0'])
token_gold_test_fix_sents   = get_sent_lists(token_gold_test_fix , fields=['token_str', 'upostag', 'biose_layer0'])
token_gold_test_fix_sents.head()

write_ncrf_file(token_gold_dev_fix_sents, 'token_gold_dev_fix', keep_pos=True)
write_ncrf_file(token_gold_dev_fix_sents, 'token_gold_dev_fix')
write_ncrf_file(token_gold_train_fix_sents, 'token_gold_train_fix', keep_pos=True)
write_ncrf_file(token_gold_train_fix_sents, 'token_gold_train_fix')
write_ncrf_file(token_gold_test_fix_sents, 'token_gold_test_fix', keep_pos=True)
write_ncrf_file(token_gold_test_fix_sents, 'token_gold_test_fix')

## TODO
1. package
1. other layers files

import os
out_folder = 'data/nested'

def write_nested_file(sents, file_name):
    suffix = '.bmes'
    with open(os.path.join(out_folder, file_name+suffix), 'w', encoding='utf8') as of:
        for sent in sents:
            for fields in sent: 
                line = ' '.join(fields)
                of.write(line+'\n')
            of.write('\n')
                    
        

fields = ['form', 'biose_layer0', 'biose_layer1', 'biose_layer2', 'biose_layer3']
a_gold_dev_sents =   bclm.get_sentences_list(gold_dev, fields=fields)
a_gold_train_sents = bclm.get_sentences_list(gold_train, fields=fields)
a_gold_test_sents =  bclm.get_sentences_list(gold_test, fields=fields)
a_gold_test_sents.head()

write_nested_file(a_gold_dev_sents,   'morph_gold_dev')
write_nested_file(a_gold_train_sents, 'morph_gold_train')
write_nested_file(a_gold_test_sents,  'morph_gold_test')

## Concat BIOSE Sentences

a_biose_concat = bclm.get_token_df(spdf, fields = ['biose_layer0', 'biose_layer1', 'biose_layer2', 'biose_layer3'])

a_token_gold_dev_concat  =  a_biose_concat[a_biose_concat.set=='dev']
a_token_gold_train_concat = a_biose_concat[a_biose_concat.set=='train']
a_token_gold_test_concat =  a_biose_concat[a_biose_concat.set=='test']

fields = ['token_str', 'biose_layer0', 'biose_layer1', 'biose_layer2', 'biose_layer3']
a_token_gold_dev_concat_sents    = bclm.get_sentences_list(a_token_gold_dev_concat  , fields=fields)
a_token_gold_train_concat_sents  = bclm.get_sentences_list(a_token_gold_train_concat, fields=fields)
a_token_gold_test_concat_sents   = bclm.get_sentences_list(a_token_gold_test_concat , fields=fields)
a_token_gold_test_concat_sents.head()

write_nested_file(a_token_gold_dev_concat_sents,   'token-multi_gold_dev')
write_nested_file(a_token_gold_train_concat_sents, 'token-multi_gold_train')
write_nested_file(a_token_gold_test_concat_sents,  'token-multi_gold_test')

## Fix BIOSE Sentences

a_biose_fix = bclm.get_token_df(spdf, fields = ['upostag'], biose=['biose_layer0', 'biose_layer1', 'biose_layer2', 'biose_layer3'])

a_biose_fix.head()

a_token_gold_dev_fix  =  a_biose_fix[a_biose_fix.set=='dev']
a_token_gold_train_fix = a_biose_fix[a_biose_fix.set=='train']
a_token_gold_test_fix =  a_biose_fix[a_biose_fix.set=='test']

fields = ['token_str', 'biose_layer0', 'biose_layer1', 'biose_layer2', 'biose_layer3']
a_token_gold_dev_fix_sents    = bclm.get_sentences_list(a_token_gold_dev_fix  , fields=fields)
a_token_gold_train_fix_sents  = bclm.get_sentences_list(a_token_gold_train_fix, fields=fields)
a_token_gold_test_fix_sents   = bclm.get_sentences_list(a_token_gold_test_fix , fields=fields)
a_token_gold_test_fix_sents.head()

write_nested_file(a_token_gold_dev_fix_sents,   'token-single_gold_dev')
write_nested_file(a_token_gold_train_fix_sents, 'token-single_gold_train')
write_nested_file(a_token_gold_test_fix_sents,  'token-single_gold_test')
