==> 32_evaluate_POS.ipynb <==
import pandas as pd

import sys
sys.path.append('/home/nlp/danb')
import bclm

bclm.evaluate_treebank_files(treebank_gold_set='dev', yap_pred_set='dev', cols=['sent_id', 'token_id', 'form'], means=True)

bclm.evaluate_treebank_files(treebank_gold_set='test', yap_pred_set='test', cols=['sent_id', 'token_id', 'form'], means=True)

df = bclm.get_token_df(bclm.read_dataframe('spmrl'), fields=['upostag', 'biose_layer0'])

train_set = set(df[df.set=='train'].upostag.tolist())
dev_set = set(df[df.set=='dev'].upostag.tolist())
test_set = set(df[df.set=='test'].upostag.tolist())


missing = (dev_set - train_set) | (test_set - train_set)

missing

dev_set - train_set

test_set - train_set

df[df.sent_id==2]

df[df.sent_id==431]

df[df.sent_id==5613]

df[df.sent_]

df[df.upostag.isin(missing)].sort_values('upostag')

train_set = set(df[df.set=='train'].biose_layer0.tolist())
dev_set = set(df[df.set=='dev'].biose_layer0.tolist())
test_set = set(df[df.set=='test'].biose_layer0.tolist())

dev_set - train_set

test_set - train_set

len(train_set)

upos_sent_count = df.groupby('upostag').sent_id.nunique()
less_than=upos_sent_count[upos_sent_count<50].index.tolist()
len(less_than)


import random 
random.sample(less_than, 124)

less_than_sents = df[df.upostag.isin(less_than)].sent_id.tolist()
df[df.sent_id.isin(less_than_sents)].sent_id.nunique()
