==> 33_bclm_basics.ipynb <==
%load_ext autoreload
%autoreload 2

%matplotlib inline

import pandas as pd
import numpy as np

import bclm

spdf = bclm.read_dataframe('spmrl')

spdf.groupby('set').sent_id.agg([min, max])

dropped = [5438, 5444, 5445, 5446, 5448, 5449, 5450, 5451, 5453, 5459]
spdf = spdf[(~spdf.sent_id.isin(dropped))]

dev = spdf[spdf.set=='dev']

spdf.head().T

## Morpheme Sentences

dev_morph_sents = bclm.get_sentences_list(dev, fields=['token_id', 'token_str', 'form', 'biose_layer0'])
dev_morph_sents

## Token-Multi Sentences

dev_multi = bclm.get_token_df(dev, fields = ['biose_layer0'])
dev_multi.head()

bclm.get_sentences_list(dev_multi, fields=['token_id', 'token_str', 'biose_layer0'])

## Token-Single Sentences

dev_single = bclm.get_token_df(dev, fields = [], biose=['biose_layer0'])
dev_single.head()

bclm.get_sentences_list(dev_single, fields=['token_id', 'token_str', 'biose_layer0'])

## UD

df = bclm.read_dataframe('ud')

df.head().T

df.dtypes

df.head().T

df.groupby('set').sent_id.agg([min, max])

df.groupby('set').sent_id.nunique()

df.to_csv('align/uddf_fixed.csv.gz', index=False, compression='gzip')
