==> 34_ud_ner_corpus.ipynb <==
%load_ext autoreload
%autoreload 2

%matplotlib inline

import pandas as pd
import numpy as np

import bclm

dropped = [5438, 5444, 5445, 5446, 5448, 5449, 5450, 5451, 5453, 5459]
spdf = bclm.read_dataframe('spmrl')

global_dropped = [spdf[spdf.sent_id==d].global_sent_id.iat[0] for d in dropped]
global_dropped

spdf.head().T

uddf = bclm.read_dataframe('ud')

uddf = uddf[(~uddf.global_sent_id.isin(global_dropped))]

from collections import OrderedDict

uddf.head().T

ud_sents = list(uddf.sent_id.unique())

## Matching sents (dev, train , test)

import re 
sent_id_re = re.compile('# sent_id = (\d+)')

import bz2
def get_sent_ids_lat(path):
    sent_ids = []
    with bz2.open(path, 'rt', encoding='utf-8') as f:
        for line in f:
            if line[0]=='#':
                sid = sent_id_re.match(line)
                if sid:
                    sent_id =  int(sid.group(1))
                sent_ids.append(sent_id)
    return sent_ids

dev_sent_ids = get_sent_ids_lat('../UL_Hebrew-HTB/he_htb-ud-dev.heblex.conllul.bz2')
train_sent_ids = get_sent_ids_lat('../UL_Hebrew-HTB/he_htb-ud-train.heblex.conllul.bz2')
test_sent_ids = get_sent_ids_lat('../UL_Hebrew-HTB/he_htb-ud-test.heblex.conllul.bz2')

            

print(f'dev: {min(dev_sent_ids)}-{max(dev_sent_ids)}')
print(f'train: {min(train_sent_ids)}-{max(train_sent_ids)}')
print(f'test: {min(test_sent_ids)}-{max(test_sent_ids)}')

uddf.groupby(['set']).sent_id.agg(['min', 'max'])

def get_set_from_sent_id(gsi):
    if gsi>=min(dev_sent_ids) and gsi<=max(dev_sent_ids):
        return 'dev'
    elif gsi>=min(train_sent_ids) and gsi<=max(train_sent_ids):
        return 'train'
    elif gsi>=min(test_sent_ids) and gsi<=max(test_sent_ids):
        return 'test'
    
uddf['ud_set'] = uddf.sent_id.apply(get_set_from_sent_id)

uddf.groupby(['ud_set']).sent_id.agg(['min', 'max'])

ud_dev = uddf[uddf.ud_set=='dev']
ud_train = uddf[uddf.ud_set=='train']
ud_test = uddf[uddf.ud_set=='test']

## Morpheme NER


import os
out_folder = 'data/ud_ner/'

def write_ncrf_file(sents, file_name, keep_pos=False, bioul=False, dummy_o=False):
    if keep_pos:
        file_name += '_pos'
    if dummy_o:
        file_name += '_dummy_o'
    if bioul:
        suffix = '.bioul'
    else:
        suffix = '.bmes'
    with open(os.path.join(out_folder, file_name+suffix), 'w', encoding='utf8') as of:
        for sent in sents:
            for fields in sent: 
                new_pos = '[POS]'+fields[1]
                line = fields[0]
                if keep_pos:
                    line += ' ' + new_pos
                if len(fields)==3:
                    tag = fields[2]
                    if bioul:
                        tag = tag.replace('E-', 'L-').replace('S-', 'U-')
                    if dummy_o:
                        tag = 'O'
                    line += ' ' + tag
                of.write(line+'\n')
            of.write('\n')
                    
        

fields = ['form', 'upostag', 'biose_layer0']
gold_dev_sents = bclm.get_sentences_list(ud_dev, fields=fields)
gold_train_sents = bclm.get_sentences_list(ud_train, fields=fields)
gold_test_sents =  bclm.get_sentences_list(ud_test, fields=fields)
gold_test_sents.head()

write_ncrf_file(gold_dev_sents,   'morph_gold_dev')
write_ncrf_file(gold_train_sents, 'morph_gold_train')
write_ncrf_file(gold_test_sents, 'morph_gold_test')

write_ncrf_file(gold_dev_sents,   'morph_gold_dev', bioul=True)
write_ncrf_file(gold_train_sents, 'morph_gold_train', bioul=True)
write_ncrf_file(gold_test_sents, 'morph_gold_test', bioul=True)

write_ncrf_file(gold_dev_sents,   'morph_gold_dev', dummy_o=True)
write_ncrf_file(gold_train_sents, 'morph_gold_train', dummy_o=True)
write_ncrf_file(gold_test_sents, 'morph_gold_test', dummy_o=True)

write_ncrf_file(gold_dev_sents,   'morph_gold_dev', dummy_o=True, bioul=True)
write_ncrf_file(gold_train_sents, 'morph_gold_train', dummy_o=True, bioul=True)
write_ncrf_file(gold_test_sents, 'morph_gold_test', dummy_o=True, bioul=True)

## Concat BIOSE Sentences

uddf['set'] = uddf['ud_set']
biose_concat = bclm.get_token_df(uddf, fields = ['biose_layer0', 'upostag'])
biose_concat.head()

token_gold_dev_concat  = biose_concat[biose_concat.set=='dev']
token_gold_train_concat = biose_concat[biose_concat.set=='train']
token_gold_test_concat = biose_concat[biose_concat.set=='test']

token_gold_dev_concat_sents    = bclm.get_sentences_list(token_gold_dev_concat  , fields=['token_str', 'upostag', 'biose_layer0'])
token_gold_train_concat_sents  = bclm.get_sentences_list(token_gold_train_concat, fields=['token_str', 'upostag', 'biose_layer0'])
token_gold_test_concat_sents   = bclm.get_sentences_list(token_gold_test_concat , fields=['token_str', 'upostag', 'biose_layer0'])
token_gold_test_concat_sents.head()

write_ncrf_file(token_gold_dev_concat_sents, 'token_gold_dev_concat')
write_ncrf_file(token_gold_train_concat_sents, 'token_gold_train_concat')
write_ncrf_file(token_gold_test_concat_sents, 'token_gold_test_concat')

write_ncrf_file(token_gold_dev_concat_sents, 'token_gold_dev_concat', bioul=True)
write_ncrf_file(token_gold_train_concat_sents, 'token_gold_train_concat', bioul=True)
write_ncrf_file(token_gold_test_concat_sents, 'token_gold_test_concat', bioul=True)

write_ncrf_file(token_gold_dev_concat_sents, 'token_gold_dev_concat', dummy_o=True)
write_ncrf_file(token_gold_train_concat_sents, 'token_gold_train_concat', dummy_o=True)
write_ncrf_file(token_gold_test_concat_sents, 'token_gold_test_concat', dummy_o=True)

write_ncrf_file(token_gold_dev_concat_sents, 'token_gold_dev_concat', dummy_o=True, bioul=True)
write_ncrf_file(token_gold_train_concat_sents, 'token_gold_train_concat', dummy_o=True, bioul=True)
write_ncrf_file(token_gold_test_concat_sents, 'token_gold_test_concat', dummy_o=True, bioul=True)

## Fix BIOSE Sentences

biose_fix = bclm.get_token_df(uddf, fields = ['upostag'], biose=['biose_layer0'])
biose_fix.head()

token_gold_dev_fix  =     biose_fix[biose_fix.set=='dev']
token_gold_train_fix = biose_fix[biose_fix.set=='train']
token_gold_test_fix =  biose_fix[biose_fix.set=='test']

token_gold_dev_fix_sents    = bclm.get_sentences_list(token_gold_dev_fix  , fields=['token_str', 'upostag', 'biose_layer0'])
token_gold_train_fix_sents  = bclm.get_sentences_list(token_gold_train_fix, fields=['token_str', 'upostag', 'biose_layer0'])
token_gold_test_fix_sents   = bclm.get_sentences_list(token_gold_test_fix , fields=['token_str', 'upostag', 'biose_layer0'])
token_gold_test_fix_sents.head()

write_ncrf_file(token_gold_dev_fix_sents, 'token_gold_dev_fix')
write_ncrf_file(token_gold_train_fix_sents, 'token_gold_train_fix')
write_ncrf_file(token_gold_test_fix_sents, 'token_gold_test_fix')

write_ncrf_file(token_gold_dev_fix_sents, 'token_gold_dev_fix', bioul=True)
write_ncrf_file(token_gold_train_fix_sents, 'token_gold_train_fix', bioul=True)
write_ncrf_file(token_gold_test_fix_sents, 'token_gold_test_fix', bioul=True)

write_ncrf_file(token_gold_dev_fix_sents, 'token_gold_dev_fix', dummy_o=True)
write_ncrf_file(token_gold_train_fix_sents, 'token_gold_train_fix', dummy_o=True)
write_ncrf_file(token_gold_test_fix_sents, 'token_gold_test_fix', dummy_o=True)

write_ncrf_file(token_gold_dev_fix_sents, 'token_gold_dev_fix', dummy_o=True, bioul=True)
write_ncrf_file(token_gold_train_fix_sents, 'token_gold_train_fix', dummy_o=True, bioul=True)
write_ncrf_file(token_gold_test_fix_sents, 'token_gold_test_fix', dummy_o=True, bioul=True)

## TODO
1. package
1. other layers files

import os
out_folder = 'data/ud_ner/nested'

def write_nested_file(sents, file_name):
    suffix = '.bmes'
    with open(os.path.join(out_folder, file_name+suffix), 'w', encoding='utf8') as of:
        for sent in sents:
            for fields in sent: 
                line = ' '.join(fields)
                of.write(line+'\n')
            of.write('\n')
                    
        

fields = ['form', 'biose_layer0', 'biose_layer1', 'biose_layer2', 'biose_layer3']
a_gold_dev_sents =   bclm.get_sentences_list(ud_dev, fields=fields)
a_gold_train_sents = bclm.get_sentences_list(ud_train, fields=fields)
a_gold_test_sents =  bclm.get_sentences_list(ud_test, fields=fields)
a_gold_test_sents.head()

write_nested_file(a_gold_dev_sents,   'morph_gold_dev')
write_nested_file(a_gold_train_sents, 'morph_gold_train')
write_nested_file(a_gold_test_sents,  'morph_gold_test')

## Concat BIOSE Sentences

a_biose_concat = bclm.get_token_df(uddf, fields = ['biose_layer0', 'biose_layer1', 'biose_layer2', 'biose_layer3'])

a_token_gold_dev_concat  =  a_biose_concat[a_biose_concat.set=='dev']
a_token_gold_train_concat = a_biose_concat[a_biose_concat.set=='train']
a_token_gold_test_concat =  a_biose_concat[a_biose_concat.set=='test']

fields = ['token_str', 'biose_layer0', 'biose_layer1', 'biose_layer2', 'biose_layer3']
a_token_gold_dev_concat_sents    = bclm.get_sentences_list(a_token_gold_dev_concat  , fields=fields)
a_token_gold_train_concat_sents  = bclm.get_sentences_list(a_token_gold_train_concat, fields=fields)
a_token_gold_test_concat_sents   = bclm.get_sentences_list(a_token_gold_test_concat , fields=fields)
a_token_gold_test_concat_sents.head()

write_nested_file(a_token_gold_dev_concat_sents,   'token-multi_gold_dev')
write_nested_file(a_token_gold_train_concat_sents, 'token-multi_gold_train')
write_nested_file(a_token_gold_test_concat_sents,  'token-multi_gold_test')

## Fix BIOSE Sentences

a_biose_fix = bclm.get_token_df(uddf, fields = ['upostag'], biose=['biose_layer0', 'biose_layer1', 'biose_layer2', 'biose_layer3'])

a_biose_fix.head()

a_token_gold_dev_fix  =  a_biose_fix[a_biose_fix.set=='dev']
a_token_gold_train_fix = a_biose_fix[a_biose_fix.set=='train']
a_token_gold_test_fix =  a_biose_fix[a_biose_fix.set=='test']

fields = ['token_str', 'biose_layer0', 'biose_layer1', 'biose_layer2', 'biose_layer3']
a_token_gold_dev_fix_sents    = bclm.get_sentences_list(a_token_gold_dev_fix  , fields=fields)
a_token_gold_train_fix_sents  = bclm.get_sentences_list(a_token_gold_train_fix, fields=fields)
a_token_gold_test_fix_sents   = bclm.get_sentences_list(a_token_gold_test_fix , fields=fields)
a_token_gold_test_fix_sents.head()

write_nested_file(a_token_gold_dev_fix_sents,   'token-single_gold_dev')
write_nested_file(a_token_gold_train_fix_sents, 'token-single_gold_train')
write_nested_file(a_token_gold_test_fix_sents,  'token-single_gold_test')
