==> 35_ab_annotations.ipynb <==
%load_ext autoreload
%autoreload 2

%matplotlib inline

import pandas as pd
import numpy as np

import bclm

dropped = [5438, 5444, 5445, 5446, 5448, 5449, 5450, 5451, 5453, 5459]
spdf = bclm.read_dataframe('spmrl')

global_dropped = [spdf[spdf.sent_id==d].global_sent_id.iat[0] for d in dropped]
global_dropped

uddf = bclm.read_dataframe('ud')

uddf = uddf[(~uddf.global_sent_id.isin(global_dropped))]

from collections import OrderedDict

ud_sents = list(uddf.sent_id.unique())

## Matching sents (dev, train , test)

import re 
sent_id_re = re.compile('# sent_id = (\d+)')

import bz2
def get_sent_ids_lat(path):
    sent_ids = []
    with bz2.open(path, 'rt', encoding='utf-8') as f:
        for line in f:
            if line[0]=='#':
                sid = sent_id_re.match(line)
                if sid:
                    sent_id =  int(sid.group(1))
                sent_ids.append(sent_id)
    return sent_ids

dev_sent_ids = get_sent_ids_lat('../UL_Hebrew-HTB/he_htb-ud-dev.heblex.conllul.bz2')
train_sent_ids = get_sent_ids_lat('../UL_Hebrew-HTB/he_htb-ud-train.heblex.conllul.bz2')
test_sent_ids = get_sent_ids_lat('../UL_Hebrew-HTB/he_htb-ud-test.heblex.conllul.bz2')

            

print(f'dev: {min(dev_sent_ids)}-{max(dev_sent_ids)}')
print(f'train: {min(train_sent_ids)}-{max(train_sent_ids)}')
print(f'test: {min(test_sent_ids)}-{max(test_sent_ids)}')

uddf.groupby(['set']).sent_id.agg(['min', 'max'])

def get_set_from_sent_id(gsi):
    if gsi>=min(dev_sent_ids) and gsi<=max(dev_sent_ids):
        return 'dev'
    elif gsi>=min(train_sent_ids) and gsi<=max(train_sent_ids):
        return 'train'
    elif gsi>=min(test_sent_ids) and gsi<=max(test_sent_ids):
        return 'test'
    
uddf['ud_set'] = uddf.sent_id.apply(get_set_from_sent_id)

uddf.groupby(['ud_set']).sent_id.agg(['min', 'max'])

import re
def get_biose(df):
    ner_bio = df.ner.copy()
    ner_bio[~ner_bio.str.contains('\[|_')] = 'S-' + ner_bio 
    ner_bio = ner_bio.replace('_', 'O')

    RE = '^(?P<type>[A-Z]+)\[(?P<num>\d+)\].*$'
    prev='XXX'
    new = []

    for x in ner_bio.tolist():
        if x.startswith('S') or x=='O':
            new.append(x)
        else:
            typ, num = re.match(RE, x).groups(0)
            if x==prev:
                new.append('I-'+typ)
            else:
                new.append('B-'+typ)

        prev = x
        
    # reverse pass to add Es
    prev='O'
    rev_pass = []
    for x in new[::-1]:
        if((prev=='O' or prev.startswith('B-')) and x.startswith('I-')):
            rev_pass.append(x.replace('I-', 'E-'))
        else:
            rev_pass.append(x)
        prev = x

    biose = rev_pass[::-1]
    
    return biose

def read_tsv3(path, add_biose=True):
    names = ['sent_tok_num', 'tok_offset', 'token',
             'FEAT_animacy', 'FEAT_aspect', 'FEAT_case', 'FEAT_definiteness', 
             'FEAT_degree', 'FEAT_gender', 'FEAT_mood', 'FEAT_negative', 'FEAT_numType',
             'FEAT_number', 'FEAT_person', 'FEAT_possessive', 'FEAT_pronType', 'FEAT_reflex', 'FEAT_tense', 
             'FEAT_transitivity', 'FEAT_value', 'FEAT_verbForm', 'FEAT_voice',
             'pos','pos_coarse',
             'ner_id', 'ner',
             'lemma',
             'surface_form',
             'dep_type', 'dep_flavor', 'dep_lex_morph_pos','dep_arc', #'EXTRA'
            ]
    df = (pd.read_csv(path, sep='\t', skiprows=10, header=None, names=names,
                      comment='#', skip_blank_lines=True, quoting=3)#na_values=['*', '_'])
          .assign(ner_type = lambda x: x.ner.str.split('[', expand=True).iloc[:,0])
          .assign(is_ner = lambda x: x.ner!='_')
          #.assign(biose = lambda x: get_biose(x))
         )
    return df


from ner_transforms import *


## fixes

df = read_tsv3('data/ab_annotations/train_485-600.conllu/dafnaa.tsv')
df.head()

def apply_fixes(name, df):
    if name=='train_485-600.conllu':
        df.loc[df.sent_tok_num=='85-15', 'ner'] = 'ORG[81]|GPE[82]'

## Read all sents for A,B

import os
from collections import defaultdict 
tators = defaultdict(list)
for folder in os.scandir('data/ab_annotations/'):
    if '.ipynb' in folder.name:
        continue
    first_sent_id = int(folder.name.split('.')[0].split('_')[1].split('-')[0])
    for file in os.scandir(folder):
        if '.ipynb' in file.name:
            continue
        print (file.path)
        tat = file.name.split('.')[0]
        df = read_tsv3(file.path)
        apply_fixes(folder.name, df)
        df[['sent_id','id']] = df.sent_tok_num.str.split('-',expand=True).astype(int)
        
        df['sent_id'] = df.sent_id + first_sent_id - 1
        df['file'] = folder.name
        layers = get_all_layers_biose(df)
        df = df.merge(layers, how='left', on=['sent_id', 'id'])
        tators[tat].append(df)
        
for tat in tators:
    tators[tat] = pd.concat(tators[tat], sort=False)
    tators[tat] = tators[tat].sort_values(by=['sent_id', 'id'])

tators['dafnaa'].tail().T

## align with UD curation output files


tators['dafnaa'].shape, tators['zefs'].shape

uddf.shape

tators['dafnaa'] = tators['dafnaa'][tators['dafnaa'].sent_id.isin(uddf.sent_id.unique())]
tators['zefs']   = tators['zefs'][tators['zefs'].sent_id.isin(uddf.sent_id.unique())]

tators['dafnaa'].shape, tators['zefs'].shape

tators['dafnaa'] = tators['dafnaa'].reset_index(drop=True)
tators['zefs']   = tators['zefs'].reset_index(drop=True)

uddf.head().T

uddf.columns

tators['dafnaa'].columns

def merge_uddf(uddf, df):
    keep_cols = ['sent_id', 'id', 'form', 'lemma', 'upostag', 'xpostag', 'feats', 'head',
       'deprel', 'deps', 'misc', 'global_sent_id', 'token_id', 'token_str',
       'token_morph_id', 'binyan', 
       'duplicate_sent_id', 'very_similar_sent_id', 'set', 'ud_set']
    tat_cols = ['ner', 'ner_type', 'is_ner', 'biose_layer0', 'biose_layer1', 'biose_layer2']
    if 'biose_layer3' in df.columns:
        tat_cols.append('biose_layer3')
    n = pd.concat([uddf[keep_cols], df[tat_cols]], axis=1)
    return n

a = merge_uddf(uddf, tators['dafnaa'])
b = merge_uddf(uddf, tators['zefs'])


a.head()

b.head()

b['biose_layer3'] = 'O'

a[['biose_layer'+str(x) for x in range(4)]] = a[['biose_layer'+str(x) for x in range(4)]].fillna('O')

b[['biose_layer'+str(x) for x in range(4)]] = b[['biose_layer'+str(x) for x in range(4)]].fillna('O')

## Create variant files for each

import os
out_folder = 'data/ud_ner/ab_annotations/'

def write_ncrf_file(sents, file_name, keep_pos=False, bioul=False, dummy_o=False):
    if keep_pos:
        file_name += '_pos'
    if dummy_o:
        file_name += '_dummy_o'
    if bioul:
        suffix = '.bioul'
    else:
        suffix = '.bmes'
    with open(os.path.join(out_folder, file_name+suffix), 'w', encoding='utf8') as of:
        for sent in sents:
            for fields in sent: 
                new_pos = '[POS]'+fields[1]
                line = fields[0]
                if keep_pos:
                    line += ' ' + new_pos
                if len(fields)==3:
                    tag = fields[2]
                    if bioul:
                        tag = tag.replace('E-', 'L-').replace('S-', 'U-')
                    if dummy_o:
                        tag = 'O'
                    line += ' ' + tag
                of.write(line+'\n')
            of.write('\n')
                    
        

a_ud_dev = a[a.ud_set=='dev']
a_ud_train = a[a.ud_set=='train']
a_ud_test = a[a.ud_set=='test']

b_ud_dev = b[b.ud_set=='dev']
b_ud_train = b[b.ud_set=='train']
b_ud_test = b[b.ud_set=='test']

fields = ['form', 'upostag', 'biose_layer0']
a_gold_dev_sents =   bclm.get_sentences_list(a_ud_dev, fields=fields)
a_gold_train_sents = bclm.get_sentences_list(a_ud_train, fields=fields)
a_gold_test_sents =  bclm.get_sentences_list(a_ud_test, fields=fields)
a_gold_test_sents.head()

fields = ['form', 'upostag', 'biose_layer0']
b_gold_dev_sents =   bclm.get_sentences_list(b_ud_dev, fields=fields)
b_gold_train_sents = bclm.get_sentences_list(b_ud_train, fields=fields)
b_gold_test_sents =  bclm.get_sentences_list(b_ud_test, fields=fields)
b_gold_test_sents.head()

write_ncrf_file(a_gold_dev_sents,   'a_morph_gold_dev')
write_ncrf_file(a_gold_train_sents, 'a_morph_gold_train')
write_ncrf_file(a_gold_test_sents,  'a_morph_gold_test')

write_ncrf_file(b_gold_dev_sents,   'b_morph_gold_dev')
write_ncrf_file(b_gold_train_sents, 'b_morph_gold_train')
write_ncrf_file(b_gold_test_sents,  'b_morph_gold_test')

## Concat BIOSE Sentences

a['set'] = a['ud_set']
b['set'] = b['ud_set']

a_biose_concat = bclm.get_token_df(a, fields = ['biose_layer0', 'upostag'])
b_biose_concat = bclm.get_token_df(b, fields = ['biose_layer0', 'upostag'])

b_biose_concat.head()

a_token_gold_dev_concat  =  a_biose_concat[a_biose_concat.set=='dev']
a_token_gold_train_concat = a_biose_concat[a_biose_concat.set=='train']
a_token_gold_test_concat =  a_biose_concat[a_biose_concat.set=='test']

b_token_gold_dev_concat  =  b_biose_concat[b_biose_concat.set=='dev']
b_token_gold_train_concat = b_biose_concat[b_biose_concat.set=='train']
b_token_gold_test_concat =  b_biose_concat[b_biose_concat.set=='test']

a_token_gold_dev_concat_sents    = bclm.get_sentences_list(a_token_gold_dev_concat  , fields=['token_str', 'upostag', 'biose_layer0'])
a_token_gold_train_concat_sents  = bclm.get_sentences_list(a_token_gold_train_concat, fields=['token_str', 'upostag', 'biose_layer0'])
a_token_gold_test_concat_sents   = bclm.get_sentences_list(a_token_gold_test_concat , fields=['token_str', 'upostag', 'biose_layer0'])
a_token_gold_test_concat_sents.head()

b_token_gold_dev_concat_sents    = bclm.get_sentences_list(b_token_gold_dev_concat  , fields=['token_str', 'upostag', 'biose_layer0'])
b_token_gold_train_concat_sents  = bclm.get_sentences_list(b_token_gold_train_concat, fields=['token_str', 'upostag', 'biose_layer0'])
b_token_gold_test_concat_sents   = bclm.get_sentences_list(b_token_gold_test_concat , fields=['token_str', 'upostag', 'biose_layer0'])
b_token_gold_test_concat_sents.head()

write_ncrf_file(a_token_gold_dev_concat_sents,   'a_token_gold_dev_concat')
write_ncrf_file(a_token_gold_train_concat_sents, 'a_token_gold_train_concat')
write_ncrf_file(a_token_gold_test_concat_sents,  'a_token_gold_test_concat')

write_ncrf_file(b_token_gold_dev_concat_sents,   'b_token_gold_dev_concat')
write_ncrf_file(b_token_gold_train_concat_sents, 'b_token_gold_train_concat')
write_ncrf_file(b_token_gold_test_concat_sents,  'b_token_gold_test_concat')

## Fix BIOSE Sentences

a_biose_fix = bclm.get_token_df(a, fields = ['upostag'], biose=['biose_layer0'])
b_biose_fix = bclm.get_token_df(b, fields = ['upostag'], biose=['biose_layer0'])

a_biose_fix.head()

a_token_gold_dev_fix  =  a_biose_fix[a_biose_fix.set=='dev']
a_token_gold_train_fix = a_biose_fix[a_biose_fix.set=='train']
a_token_gold_test_fix =  a_biose_fix[a_biose_fix.set=='test']

b_token_gold_dev_fix  =  b_biose_fix[b_biose_fix.set=='dev']
b_token_gold_train_fix = b_biose_fix[b_biose_fix.set=='train']
b_token_gold_test_fix =  b_biose_fix[b_biose_fix.set=='test']

a_token_gold_dev_fix_sents    = bclm.get_sentences_list(a_token_gold_dev_fix  , fields=['token_str', 'upostag', 'biose_layer0'])
a_token_gold_train_fix_sents  = bclm.get_sentences_list(a_token_gold_train_fix, fields=['token_str', 'upostag', 'biose_layer0'])
a_token_gold_test_fix_sents   = bclm.get_sentences_list(a_token_gold_test_fix , fields=['token_str', 'upostag', 'biose_layer0'])
a_token_gold_test_fix_sents.head()

b_token_gold_dev_fix_sents    = bclm.get_sentences_list(b_token_gold_dev_fix  , fields=['token_str', 'upostag', 'biose_layer0'])
b_token_gold_train_fix_sents  = bclm.get_sentences_list(b_token_gold_train_fix, fields=['token_str', 'upostag', 'biose_layer0'])
b_token_gold_test_fix_sents   = bclm.get_sentences_list(b_token_gold_test_fix , fields=['token_str', 'upostag', 'biose_layer0'])
b_token_gold_test_fix_sents.head()

write_ncrf_file(a_token_gold_dev_fix_sents,   'a_token_gold_dev_fix')
write_ncrf_file(a_token_gold_train_fix_sents, 'a_token_gold_train_fix')
write_ncrf_file(a_token_gold_test_fix_sents,  'a_token_gold_test_fix')

write_ncrf_file(b_token_gold_dev_fix_sents,   'b_token_gold_dev_fix')
write_ncrf_file(b_token_gold_train_fix_sents, 'b_token_gold_train_fix')
write_ncrf_file(b_token_gold_test_fix_sents,  'b_token_gold_test_fix')

## TODO
1. package
1. other layers files

import os
out_folder = 'data/ud_ner/ab_annotations/nested'

def write_nested_file(sents, file_name):
    suffix = '.bmes'
    with open(os.path.join(out_folder, file_name+suffix), 'w', encoding='utf8') as of:
        for sent in sents:
            for fields in sent: 
                line = ' '.join(fields)
                of.write(line+'\n')
            of.write('\n')
                    
        

fields = ['form', 'biose_layer0', 'biose_layer1', 'biose_layer2', 'biose_layer3']
a_gold_dev_sents =   bclm.get_sentences_list(a_ud_dev, fields=fields)
a_gold_train_sents = bclm.get_sentences_list(a_ud_train, fields=fields)
a_gold_test_sents =  bclm.get_sentences_list(a_ud_test, fields=fields)
a_gold_test_sents.head()

fields = ['form', 'biose_layer0', 'biose_layer1', 'biose_layer2', 'biose_layer3']
b_gold_dev_sents =   bclm.get_sentences_list(b_ud_dev, fields=fields)
b_gold_train_sents = bclm.get_sentences_list(b_ud_train, fields=fields)
b_gold_test_sents =  bclm.get_sentences_list(b_ud_test, fields=fields)
b_gold_test_sents.head()

write_nested_file(a_gold_dev_sents,   'a_morph_gold_dev')
write_nested_file(a_gold_train_sents, 'a_morph_gold_train')
write_nested_file(a_gold_test_sents,  'a_morph_gold_test')

write_nested_file(b_gold_dev_sents,   'b_morph_gold_dev')
write_nested_file(b_gold_train_sents, 'b_morph_gold_train')
write_nested_file(b_gold_test_sents,  'b_morph_gold_test')

## Concat BIOSE Sentences

a['set'] = a['ud_set']
b['set'] = b['ud_set']

a_biose_concat = bclm.get_token_df(a, fields = ['biose_layer0', 'biose_layer1', 'biose_layer2', 'biose_layer3'])
b_biose_concat = bclm.get_token_df(b, fields = ['biose_layer0', 'biose_layer1', 'biose_layer2', 'biose_layer3'])

b_biose_concat.head()

a_token_gold_dev_concat  =  a_biose_concat[a_biose_concat.set=='dev']
a_token_gold_train_concat = a_biose_concat[a_biose_concat.set=='train']
a_token_gold_test_concat =  a_biose_concat[a_biose_concat.set=='test']

b_token_gold_dev_concat  =  b_biose_concat[b_biose_concat.set=='dev']
b_token_gold_train_concat = b_biose_concat[b_biose_concat.set=='train']
b_token_gold_test_concat =  b_biose_concat[b_biose_concat.set=='test']

fields = ['token_str', 'biose_layer0', 'biose_layer1', 'biose_layer2', 'biose_layer3']
a_token_gold_dev_concat_sents    = bclm.get_sentences_list(a_token_gold_dev_concat  , fields=fields)
a_token_gold_train_concat_sents  = bclm.get_sentences_list(a_token_gold_train_concat, fields=fields)
a_token_gold_test_concat_sents   = bclm.get_sentences_list(a_token_gold_test_concat , fields=fields)
a_token_gold_test_concat_sents.head()

fields = ['token_str', 'biose_layer0', 'biose_layer1', 'biose_layer2', 'biose_layer3']
b_token_gold_dev_concat_sents    = bclm.get_sentences_list(b_token_gold_dev_concat  , fields=fields)
b_token_gold_train_concat_sents  = bclm.get_sentences_list(b_token_gold_train_concat, fields=fields)
b_token_gold_test_concat_sents   = bclm.get_sentences_list(b_token_gold_test_concat , fields=fields)
b_token_gold_test_concat_sents.head()

write_nested_file(a_token_gold_dev_concat_sents,   'a_token-multi_gold_dev')
write_nested_file(a_token_gold_train_concat_sents, 'a_token-multi_gold_train')
write_nested_file(a_token_gold_test_concat_sents,  'a_token-multi_gold_test')

write_nested_file(b_token_gold_dev_concat_sents,   'b_token-multi_gold_dev')
write_nested_file(b_token_gold_train_concat_sents, 'b_token-multi_gold_train')
write_nested_file(b_token_gold_test_concat_sents,  'b_token-multi_gold_test')

## Fix BIOSE Sentences

a_biose_fix = bclm.get_token_df(a, fields = ['upostag'], biose=['biose_layer0', 'biose_layer1', 'biose_layer2', 'biose_layer3'])

b_biose_fix = bclm.get_token_df(b, fields = ['upostag'], biose=['biose_layer0', 'biose_layer1', 'biose_layer2', 'biose_layer3'])

b_biose_fix.head()

a_token_gold_dev_fix  =  a_biose_fix[a_biose_fix.set=='dev']
a_token_gold_train_fix = a_biose_fix[a_biose_fix.set=='train']
a_token_gold_test_fix =  a_biose_fix[a_biose_fix.set=='test']

b_token_gold_dev_fix  =  b_biose_fix[b_biose_fix.set=='dev']
b_token_gold_train_fix = b_biose_fix[b_biose_fix.set=='train']
b_token_gold_test_fix =  b_biose_fix[b_biose_fix.set=='test']

fields = ['token_str', 'biose_layer0', 'biose_layer1', 'biose_layer2', 'biose_layer3']
a_token_gold_dev_fix_sents    = bclm.get_sentences_list(a_token_gold_dev_fix  , fields=fields)
a_token_gold_train_fix_sents  = bclm.get_sentences_list(a_token_gold_train_fix, fields=fields)
a_token_gold_test_fix_sents   = bclm.get_sentences_list(a_token_gold_test_fix , fields=fields)
a_token_gold_test_fix_sents.head()

fields = ['token_str', 'biose_layer0', 'biose_layer1', 'biose_layer2', 'biose_layer3']
b_token_gold_dev_fix_sents    = bclm.get_sentences_list(b_token_gold_dev_fix  , fields=fields)
b_token_gold_train_fix_sents  = bclm.get_sentences_list(b_token_gold_train_fix, fields=fields)
b_token_gold_test_fix_sents   = bclm.get_sentences_list(b_token_gold_test_fix , fields=fields)
b_token_gold_test_fix_sents.head()

write_nested_file(a_token_gold_dev_fix_sents,   'a_token-single_gold_dev')
write_nested_file(a_token_gold_train_fix_sents, 'a_token-single_gold_train')
write_nested_file(a_token_gold_test_fix_sents,  'a_token-single_gold_test')

write_nested_file(b_token_gold_dev_fix_sents,   'b_token-single_gold_dev')
write_nested_file(b_token_gold_train_fix_sents, 'b_token-single_gold_train')
write_nested_file(b_token_gold_test_fix_sents,  'b_token-single_gold_test')
