==> 36_ud_json_files.ipynb <==
install https://github.com/cjer/bclm  
install https://github.com/EmilStenstrom/conllu

%load_ext autoreload
%autoreload 2

%matplotlib inline

import pandas as pd
import numpy as np

import bclm

dropped = [5438, 5444, 5445, 5446, 5448, 5449, 5450, 5451, 5453, 5459]
spdf = bclm.read_dataframe('spmrl')

global_dropped = [spdf[spdf.sent_id==d].global_sent_id.iat[0] for d in dropped]
global_dropped

uddf = bclm.read_dataframe('ud')

uddf = uddf[(~uddf.global_sent_id.isin(global_dropped))]

spdf = spdf[(~spdf.global_sent_id.isin(global_dropped))]

from collections import OrderedDict

ud_sents = list(uddf.sent_id.unique())

def get_set_from_sent_id(gsi):
    if gsi>=1 and gsi<=484:
        return 'dev'
    elif gsi>=485 and gsi<=5725:
        return 'train'
    elif gsi>=5726 and gsi<=6216:
        return 'test'
    
uddf['ud_set'] = uddf.sent_id.apply(get_set_from_sent_id)

uddf.groupby(['ud_set']).sent_id.agg(['min', 'max'])

uddf['set'] = uddf['ud_set']

## TODO:
1. Add text
1. calculate roots and edges
1. Calculate NER start/stop positions
1. Add NER for all layers
1. Create dict from df sentence
1. Write jsonl
1. NOT NOW: add offsets

### Add text

from conllu import parse 
with open('align/ud_fixed.conllu', 'r', encoding='utf8') as f:
    ud_conllu_sents = parse(f.read())

ud_conllu_sents[1].metadata

ud_texts = {int(s.metadata['sent_id']): s.metadata['text'] for s in ud_conllu_sents}
len(ud_texts)

uddf['text'] = uddf.sent_id.map(ud_texts)
uddf.head().T

from conllu import parse 
with open('align/spmrl_fixed.conllu', 'r', encoding='utf8') as f:
    sp_conllu_sents = parse(f.read())

sp_conllu_sents[1].metadata

sp_texts = {int(s.metadata['sent_id']): s.metadata['text_from_ud'] for s in sp_conllu_sents}
len(sp_texts)

spdf['text'] = spdf.sent_id.map(sp_texts)
spdf.head().T

def get_edges(df):
    return (df
            .rename(columns={'head': 'source', 
                             'id': 'destination', 
                             'deprel': 'relation'})[['source', 'destination', 'relation']]
            .to_dict('records'))

def get_roots(df):
    return [int(x) for x in list(df[(df['head']==0) | (df.deprel.str.lower()=='root')].id.unique())]

def get_overlapping_entities(df, col='biose_layer0', priority=0):
    mentions = []
    current_mention= None
    for i, r in df.iterrows():
        bio = r[col][0]
        if bio!='O':
            cat = r[col].split('-')[1]
        else:
            cat = '-'
        if bio=='S':
            mentions.append({'first': int(r['id']),
                             'last': int(r['id']),
                             'label': cat, 
                             'priority': priority})
            current_mention= None
        if bio=='B':
            current_mention = {'first': int(r['id']), 
                               'last': int(r['id']), 
                               'label': cat,
                               'priority': priority}

#         if bio=='I' and current_mention is not None:
#             pass
        
        if bio=='E' and current_mention is not None:
            current_mention['last'] = int(r['id'])
            mentions.append(current_mention)
            current_mention= None
        if bio=='O':
            current_mention = None
    return mentions
    
def get_sentence_dict(df):
    df = df.sort_values('id')
    d = {}
    d['id'] = int(df.sent_id.iat[0])
    d['text'] = df.text.iat[0]
    s = {}
    s['words'] = df.form.tolist()
    s['raw'] = s['words']
    s['startOffsets'] = []
    s['endOffsets'] = []
    s['tags'] = df.upostag.tolist()
    s['lemmas'] = df.lemma.tolist()
    s['chunks'] = []
    edges = get_edges(df)
    roots = get_roots(df)
    universal_basic = {'edges': edges, 'roots': roots}
    s['graphs'] = {'universal-basic': universal_basic,
                   'universal_enhanced': {'edges': [], 'roots': []}}
    s['metadata'] =  {
                'sentParagraphOffsets': []
            },
    ents = (get_overlapping_entities(df, 'biose_layer0', 0) 
            + get_overlapping_entities(df, 'biose_layer1', 1) 
            + get_overlapping_entities(df, 'biose_layer2', 2) 
            + get_overlapping_entities(df, 'biose_layer3', 3))
    s['overlappingEntities'] = ents
    d['sentences'] = [s,]
    d['metadata'] = {'paragraph_': d['text'], 
                     'section': df.set.iat[0]}
    d['gold_relations']= [],
    d['gold_events']= [],
    d['corefChains']= {'rawMentions': []}
    return d

get_sentence_dict(uddf[uddf.sent_id==1])
    
    

get_sentence_dict(spdf[spdf.sent_id==1])
    
    

all_dicts = [get_sentence_dict(uddf[uddf.sent_id==i]) for i in sorted(list(uddf.sent_id.unique()))]

import jsonlines

with jsonlines.open('data/hebtb_ud.ann.jsonl', mode='w') as writer:
    for d in all_dicts:
        writer.write(d)

sp_all_dicts = [get_sentence_dict(spdf[spdf.sent_id==i]) for i in sorted(list(spdf.sent_id.unique()))]

import jsonlines

with jsonlines.open('data/hebtb_spmrl.ann.jsonl', mode='w') as writer:
    for d in sp_all_dicts:
        writer.write(d)
