==> 37_pilot_annotations.ipynb <==
%load_ext autoreload
%autoreload 2

%matplotlib inline

import pandas as pd
import numpy as np

import bclm

dropped = [5438, 5444, 5445, 5446, 5448, 5449, 5450, 5451, 5453, 5459]
spdf = bclm.read_dataframe('spmrl', subset='dev')

global_dropped = [spdf[spdf.sent_id==d].global_sent_id.iat[0] for d in dropped]
global_dropped

uddf = bclm.read_dataframe('ud', subset='dev')

uddf = uddf[(~uddf.global_sent_id.isin(global_dropped))]

from collections import OrderedDict

ud_sents = list(uddf.sent_id.unique())

## Matching sents (dev, train , test)

import re 
sent_id_re = re.compile('# sent_id = (\d+)')

import bz2
def get_sent_ids_lat(path):
    sent_ids = []
    with bz2.open(path, 'rt', encoding='utf-8') as f:
        for line in f:
            if line[0]=='#':
                sid = sent_id_re.match(line)
                if sid:
                    sent_id =  int(sid.group(1))
                sent_ids.append(sent_id)
    return sent_ids

dev_sent_ids = get_sent_ids_lat('../UL_Hebrew-HTB/he_htb-ud-dev.heblex.conllul.bz2')
            

print(f'dev: {min(dev_sent_ids)}-{max(dev_sent_ids)}')

uddf.groupby(['set']).sent_id.agg(['min', 'max'])

def get_set_from_sent_id(gsi):
    if gsi>=min(dev_sent_ids) and gsi<=max(dev_sent_ids):
        return 'dev'
    
uddf['ud_set'] = uddf.sent_id.apply(get_set_from_sent_id)

uddf.groupby(['ud_set']).sent_id.agg(['min', 'max'])

import re
def get_biose(df):
    ner_bio = df.ner.copy()
    ner_bio[~ner_bio.str.contains('\[|_')] = 'S-' + ner_bio 
    ner_bio = ner_bio.replace('_', 'O')

    RE = '^(?P<type>[A-Z]+)\[(?P<num>\d+)\].*$'
    prev='XXX'
    new = []

    for x in ner_bio.tolist():
        if x.startswith('S') or x=='O':
            new.append(x)
        else:
            typ, num = re.match(RE, x).groups(0)
            if x==prev:
                new.append('I-'+typ)
            else:
                new.append('B-'+typ)

        prev = x
        
    # reverse pass to add Es
    prev='O'
    rev_pass = []
    for x in new[::-1]:
        if((prev=='O' or prev.startswith('B-')) and x.startswith('I-')):
            rev_pass.append(x.replace('I-', 'E-'))
        else:
            rev_pass.append(x)
        prev = x

    biose = rev_pass[::-1]
    
    return biose

def read_tsv3(path, add_biose=True):
    names = ['sent_tok_num', 'tok_offset', 'token',
              'FEAT_gender', 'FEAT_number', 'FEAT_case', 'FEAT_degree', 'FEAT_transitivity',
              'FEAT_tense', 'FEAT_mood', 'FEAT_voice', 'FEAT_definiteness', 'FEAT_value', 'FEAT_person', 'FEAT_aspect',
              'pos',
              'ner',
              'lemma',
              'surface_form',
              'dep_type', 'dep_flavor', 'dep_lex_morph_pos', 'dep_arc', 'EXTRA'
            ]
    df = (pd.read_csv(path, sep='\t', skiprows=10, header=None, names=names,
                      comment='#', skip_blank_lines=True, quoting=3)#na_values=['*', '_'])
          .assign(ner_type = lambda x: x.ner.str.split('[', expand=True).iloc[:,0])
          .assign(is_ner = lambda x: x.ner!='_')
          #.assign(biose = lambda x: get_biose(x))
         )
    return df


from ner_transforms import *


## fixes

df = read_tsv3('data/pilot_annotations/group1/anatb/dev_1-100.tsv')
df.head()

def apply_fixes(name, df):
    if name=='train_485-600.conllu':
        df.loc[df.sent_tok_num=='85-15', 'ner'] = 'ORG[81]|GPE[82]'

## Read all sents for A,B

import os
from collections import defaultdict 
groups = {}
for folder in os.scandir('data/pilot_annotations/'):
    if '.ipynb' in folder.name:
        continue
    group = folder.name
    groups[group] = defaultdict(list)
    for tat_folder in os.scandir(folder):
        if '.ipynb' in tat_folder.name:
            continue
        tat = tat_folder.name
        for file in os.scandir(tat_folder):
            if '.ipynb' in file.name:
                continue
            print (file.path)
            first_sent_id = int(file.name.split('.')[0].split('_')[1].split('-')[0])
            df = read_tsv3(file.path)
            #apply_fixes(folder.name, df)
            df[['sent_id','id']] = df.sent_tok_num.str.split('-',expand=True).astype(int)

            df['sent_id'] = df.sent_id + first_sent_id - 1
            df['file'] = folder.name
            layers = get_all_layers_biose(df)
            df = df.merge(layers, how='left', on=['sent_id', 'id'])
            groups[group][tat].append(df)
        
for group in groups:
    for tat in groups[group]:
        groups[group][tat] = pd.concat(groups[group][tat], sort=False)
        groups[group][tat] = groups[group][tat].sort_values(by=['sent_id', 'id'])

groups['group1']['anatb'].tail().T

## align with UD curation output files


for group in groups:
    for tat in groups[group]:
        print(group, tat, groups[group][tat].shape)

uddf = uddf[uddf.sent_id<=400]
uddf.shape

sent_diff = groups['group1']['anatb'].groupby('sent_id').size()[groups['group1']['anatb'].groupby('sent_id').size()!=uddf.groupby('sent_id').size()]
sent_diff

for group, tators in groups.items():
    for tat, annos in tators.items():
        print(tat)
        #print(annos.loc[(annos.sent_id.isin(sent_diff.index)) & (annos.ner!='_'), ['sent_id', 'id', 'token', 'ner']])
        print(annos.loc[(annos.sent_id.isin(sent_diff.index)) & (annos.ner!='_'), :].sent_tok_num.unique())

uddf[uddf.sent_id.isin(sent_diff.index[3:4])]

groups['group1']['anatb'][groups['group1']['anatb'].sent_id.isin(sent_diff.index[3:4])]

for group in groups:
    for tat in groups[group]:
         groups[group][tat] =  groups[group][tat].reset_index(drop=True)

uddf.columns

uddf = uddf[uddf.sent_id<=400]
uddf.shape

keep_cols = ['sent_id', 'id', 'form', 'lemma', 'upostag', 'xpostag', 'feats', 'head',
       'deprel', 'deps', 'misc', 'global_sent_id', 'token_id', 'token_str',
       'token_morph_id', 'binyan', 
       'duplicate_sent_id', 'very_similar_sent_id', 'set', 'ud_set']
def merge_uddf(uddf, df):
    keep_cols = ['sent_id', 'id', 'form', 'lemma', 'upostag', 'xpostag', 'feats', 'head',
       'deprel', 'deps', 'misc', 'global_sent_id', 'token_id', 'token_str',
       'token_morph_id', 'binyan', 
       'duplicate_sent_id', 'very_similar_sent_id', 'set', 'ud_set']
    tat_cols = ['ner', 'ner_type', 'is_ner', 'biose_layer0']
    for i in range(4):
        if 'biose_layer'+str(i+1) in df.columns:
            tat_cols.append('biose_layer'+str(i+1))
    n = pd.concat([uddf[keep_cols], df[tat_cols]], axis=1)
    return n, keep_cols+tat_cols

uddf_tofix = uddf[(uddf.sent_id.isin(sent_diff.index))][keep_cols]

uddf_tofix.shape

uddf_ok = uddf[~(uddf.sent_id.isin(sent_diff.index))].reset_index(drop=True)

merged = {}
for group, tators in groups.items():
    print(group)
    merged[group] = {}
    for tat, annos in tators.items():
        print(tat)
        annos_ok = annos[~(annos.sent_id.isin(sent_diff.index))].reset_index(drop=True)
        a, all_cols = merge_uddf(uddf_ok, annos_ok)
        biose_cols = [c for c in all_cols if 'biose_layer' in c]
        a = a[all_cols]
        annos_tofix = annos[(annos.sent_id.isin(sent_diff.index)) 
                         & (annos.ner!='_')][['sent_id', 'token', 'ner', 'ner_type', 'is_ner'] +biose_cols].rename(columns={'token': 'form'})
        a_missing = uddf_tofix.merge(annos_tofix, how='left', on=['sent_id', 'form'])[all_cols]
        a = pd.concat([a, a_missing])
        a = a.sort_values(['sent_id', 'id']).reset_index(drop=True)
        a = a[a.sent_id.notna()]
        a[['ner', 'ner_type']] = a[['ner', 'ner_type']].fillna('_')
        a['is_ner'] = a['is_ner'].fillna(False)
        for i in range(4):
            if 'biose_layer'+str(i) in a.columns:
                a['biose_layer'+str(i)] = a['biose_layer'+str(i)].fillna('O')
            else:
                a['biose_layer'+str(i)] = 'O'
        merged[group][tat] = a
        

merged['group1']['anatb'].shape

## Create variant files for each

aliases = {'dafnaa': 'a', 
           'zefs': 'b'}
i = 0
for group in groups:
    for tat in groups[group]:
        if tat not in aliases:
            aliases[tat] = chr(ord('c')+i)
            i+=1
aliases

import os
out_folder = 'data/ud_ner/pilot_annotations/'

def write_ncrf_file(sents, file_name, keep_pos=False, bioul=False, dummy_o=False):
    if keep_pos:
        file_name += '_pos'
    if dummy_o:
        file_name += '_dummy_o'
    if bioul:
        suffix = '.bioul'
    else:
        suffix = '.bmes'
    with open(os.path.join(out_folder, file_name+suffix), 'w', encoding='utf8') as of:
        for sent in sents:
            for fields in sent: 
                new_pos = '[POS]'+fields[1]
                line = fields[0]
                if keep_pos:
                    line += ' ' + new_pos
                if len(fields)==3:
                    tag = fields[2]
                    if bioul:
                        tag = tag.replace('E-', 'L-').replace('S-', 'U-')
                    if dummy_o:
                        tag = 'O'
                    line += ' ' + tag
                of.write(line+'\n')
            of.write('\n')
                    
        

import os
nested_out_folder = 'data/ud_ner/pilot_annotations/nested'

def write_nested_file(sents, file_name):
    suffix = '.bmes'
    with open(os.path.join(nested_out_folder, file_name+suffix), 'w', encoding='utf8') as of:
        for sent in sents:
            for fields in sent: 
                line = ' '.join(fields)
                of.write(line+'\n')
            of.write('\n')
            

1==1

for group in merged:
    for name, a in merged[group].items():
        alias = aliases[name]
        
        #morph
        sents = bclm.get_sentences_list(a, fields=['form', 'upostag', 'biose_layer0'])
        write_ncrf_file(sents, f'{group}_{alias}_morph_dev_1-400')
        
        #token-multi
        a_biose_concat = bclm.get_token_df(a, fields = ['biose_layer0', 'upostag'])
        sents = bclm.get_sentences_list(a_biose_concat  , fields=['token_str', 'upostag', 'biose_layer0'])
        write_ncrf_file(sents, f'{group}_{alias}_token-multi_dev_1-400')

        #token-single
        a_biose_fix = bclm.get_token_df(a, fields = ['upostag'], biose=['biose_layer0'])
        sents = bclm.get_sentences_list(a_biose_fix  , fields=['token_str', 'upostag', 'biose_layer0'])
        write_ncrf_file(sents, f'{group}_{alias}_token-single_dev_1-400')
        
        # === nested ===
        #morph
        fields = ['form', 'biose_layer0', 'biose_layer1', 'biose_layer2', 'biose_layer3']
        sents =   bclm.get_sentences_list(a, fields=fields)
        write_nested_file(sents, f'{group}_{alias}_morph_dev_1-400')
        
        #token-multi
        fields = ['token_str', 'biose_layer0', 'biose_layer1', 'biose_layer2', 'biose_layer3']
        a_biose_concat = bclm.get_token_df(a, fields = fields[1:])
        sents = bclm.get_sentences_list(a_biose_concat  , fields=fields)
        write_nested_file(sents, f'{group}_{alias}_token-multi_dev_1-400')
        
        #token-single
        a_biose_fix = bclm.get_token_df(a, fields = ['upostag'], biose=fields[1:])
        sents = bclm.get_sentences_list(a_biose_fix  , fields=fields)
        write_nested_file(sents, f'{group}_{alias}_token-single_dev_1-400')
